LayoutTests/http/tests/contentextensions/preconnect-blocked.html

This page preconnects a Document.<br>
The preconnect should be blocked by the content extension filter<br>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
}
</script>
<link rel="preconnect" href="resources/should-not-load.html">
