LayoutTests/http/tests/contentextensions/preconnect-blocked-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/preconnect-blocked.html from loading a resource from http://127.0.0.1:8000/contentextensions/resources/should-not-load.html
This page preconnects a Document.
The preconnect should be blocked by the content extension filter
